Create 62-Array of Structure
Why is getchar() needed here?

When you use scanf("%d", &s[i].roll_no);, it reads the integer input, but after entering the integer, when you press Enter, a newline character \n remains in the input buffer. If this newline is not consumed, the next scanf("%c", &s[i].grade); call will read the newline character instead of the intended input for grade.

By adding getchar(), the newline is removed from the buffer before reading the grade. Without it, the program would not behave as expected when trying to read the character input for grade.